Refactored redux implementation.
