+++ actions/authActions/index.js
@@ -1,0 +1,5 @@
+import * as actionTypes from '../../actionTypes';
+
+export const tokenExpired = () => ({
+    type: actionTypes.TOKEN_EXPIRED
+});
