+++ actions/index.js
@@ -1,9 +1,39 @@
-import * as actionTypes from '../actionTypes';
-export * from './postsActions';
-export * from './usersActions';
-export * from './authActions';
-export * from './currentUserActions';
-export * from './errorActions';
-export * from './successActions';
-export * from './commentActions';
+// import * as actionTypes from '../actionTypes';
+// export * from './postsActions';
+// export * from './usersActions';
+// export * from './authActions';
+// export * from './currentUserActions';
+// export * from './errorActions';
+// export * from './successActions';
+// export * from './commentActions';
+// export * from './documentActions';
+
+export * from './authActions/getPasswordResetEmail';
+export * from './authActions/setNewPassword';
+export * from './authActions/signIn';
+export * from './authActions/signOut';
+export * from './authActions/signUp';
+
+export * from './commentActions/deleteComment';
+export * from './commentActions/editComment';
+export * from './commentActions/replyToComment';
+export * from './commentActions/replyToPost';
+
+export * from './currentUserActions/fetchCurrentUser';
+export * from './currentUserActions/followUser';
+export * from './currentUserActions/giveKudos';
+export * from './currentUserActions/removeKudos';
+export * from './currentUserActions/unfollowUser';
+export * from './currentUserActions/updateUserAvatar';
+export * from './currentUserActions/updateUserDetails';
+export * from './currentUserActions/updateUserPassword';
+
+export * from './postsActions/createPost';
+export * from './postsActions/editPost';
+export * from './postsActions/fetchCategoriesPosts';
+export * from './postsActions/fetchPost';
+export * from './postsActions/fetchPosts';
+
+export * from './usersActions/fetchUser';
+
 export * from './documentActions';
