+++ pages/edit-post.js
@@ -1,7 +1,7 @@
 import withRedux from 'next-redux-wrapper';
 import makeStore from '../store';
 import { fetchCurrentUser, fetchPost } from '../actions';
-import { fetchCurrentUserIfNeeded, cleanseErrorsAndSuccesses } from '../utils';
+import { fetchCurrentUserIfNeeded } from '../utils';
 import Header from '../components/Header';
 import EditArticleForm from '../components/EditArticleForm';
 import ArticleEditor from '../components/ArticleEditor';
@@ -27,7 +27,6 @@ const editPost = props => {
 
 editPost.getInitialProps = async ({store, isServer, req, pathname, query}) => {
     const currentState = store.getState();
-    cleanseErrorsAndSuccesses(store);
     const currentUser = fetchCurrentUserIfNeeded(currentState, store);
     const post = store.dispatch(fetchPost(query.post));
     await Promise.all([currentUser, post]);
