+++ pages/forgot.js
@@ -1,7 +1,7 @@
 import withRedux from 'next-redux-wrapper';
 import makeStore from '../store';
 import { fetchCurrentUser, cleanseErrors } from '../actions';
-import { fetchCurrentUserIfNeeded, cleanseErrorsAndSuccesses } from '../utils';
+import { fetchCurrentUserIfNeeded } from '../utils';
 import Header from '../components/Header';
 import PasswordResetEmailReqForm from '../components/PasswordResetEmailReqForm';
 
@@ -14,7 +14,6 @@ const forgot = props => (
 
 forgot.getInitialProps = async ({store, isServer, req, pathname, query}) => {
     const currentState = store.getState();
-    cleanseErrorsAndSuccesses(store);
     await fetchCurrentUserIfNeeded(currentState, store);
     return;
 };
