+++ pages/reset.js
@@ -1,7 +1,7 @@
 import withRedux from 'next-redux-wrapper';
 import makeStore from '../store';
 import { fetchCurrentUser } from '../actions';
-import { fetchCurrentUserIfNeeded, cleanseErrorsAndSuccesses } from '../utils';
+import { fetchCurrentUserIfNeeded } from '../utils';
 import Header from '../components/Header';
 import PasswordResetForm from '../components/PasswordResetForm';
 
@@ -14,7 +14,6 @@ const reset = props => (
 
 reset.getInitialProps = async ({store, isServer, req, pathname, query}) => {
     const currentState = store.getState();
-    cleanseErrorsAndSuccesses(store);
     await fetchCurrentUserIfNeeded(currentState, store);
     return {
         resetPasswordToken: query.resetPasswordToken
