+++ pages/user.js
@@ -1,7 +1,7 @@
 import withRedux from 'next-redux-wrapper';
 import makeStore from '../store';
 import { fetchUser, fetchCurrentUser } from '../actions';
-import { fetchCurrentUserIfNeeded, fetchUserIfNeeded, cleanseErrorsAndSuccesses } from '../utils';
+import { fetchCurrentUserIfNeeded, fetchUserIfNeeded } from '../utils';
 import { Wrapper } from '../components/Layout';
 import Header from '../components/Header';
 import UserProfileHeader from '../components/UserProfileHeader';
@@ -19,7 +19,6 @@ const user = props => (
 
 user.getInitialProps = async ({store, isServer, req, pathname, query}) => {
     const currentState = store.getState();
-    cleanseErrorsAndSuccesses(store);
     const currentUser = fetchCurrentUserIfNeeded(currentState, store);
     const user = fetchUserIfNeeded(currentState, store, query.user);
     await Promise.all([currentUser, user]);
