--- utils/cleanseErrorsAndSuccesses.js
@@ -1,6 +1,0 @@
-import { cleanseErrors, cleanseSuccesses } from '../actions';
-
-export const cleanseErrorsAndSuccesses = (store) => {
-    store.dispatch(cleanseErrors());
-    store.dispatch(cleanseSuccesses());
-};
